Move containerservice/2019-02-01 to preview since this should be a preview package
