+++ services/preview/containerservice/mgmt/2019-02-01-preview/containerservice/operations.go
